Player View, Trashcan Collision, and Attempt at Shader
So I changed the view to focus down on Bob and follow him around the map, I made him stop when he collides with a trashcan (that isn't cursed, haven't figured out the "cursing" mechanic yet) and downloaded the shader package and followed the directions on the video to implement it into our game. So far the lighting changes depending on how close Bob is to the light source I placed in the Instances layer, but the trashcans don't cast any shadows, and I haven't figured out how to delete the unnecessary light sources and have Bob be the sole light source. I'll work more on this later.