Background, Shader, and Trashcan Collision
I tiled the background with the water tile from the first workshop (we'll change it when we get the real one from Ilan). I also downloaded and implemented the shader with the video for guidance, though I will go back and fix it later. Finally, I put trashcans throughout the room and made it so Bob successfully stops when he collides with them.